+++ Security/OSX/libsecurity_ssl/regressions/SecureTransportTests/STLegacyTests+sni.m
@@ -77,7 +77,7 @@ static OSStatus SocketWrite(SSLConnectionRef h, const void *data, size_t *length
         ssize_t ret;
         do {
             hexdump(ptr, len);
-            ret = write((int)h, ptr, len);
+            ret = write((int)(intptr_t)h, ptr, len);
         } while ((ret < 0) && (errno == EAGAIN || errno == EINTR));
         if (ret > 0) {
             len -= ret;
@@ -99,13 +99,13 @@ static OSStatus SocketRead(SSLConnectionRef h, void *data, size_t *length)
     do {
         ssize_t ret;
         do {
-            ret = read((int)h, ptr, len);
+            ret = read((int)(intptr_t)h, ptr, len);
         } while ((ret < 0) && (errno == EAGAIN || errno == EINTR));
         if (ret > 0) {
             len -= ret;
             ptr += ret;
         } else {
-            printf("read error(%d): ret=%zd, errno=%d\n", (int)h, ret, errno);
+            printf("read error(%" PRIdPTR "): ret=%zd, errno=%d\n", (intptr_t)h, ret, errno);
             return -errno;
         }
     } while (len > 0);
