+++ Security/OSX/libsecurity_ssl/regressions/ssl-39-echo.c
@@ -421,7 +421,7 @@ static OSStatus SocketWrite(SSLConnectionRef conn, const void *data, size_t *len
         ssize_t ret;
         do {
             hexdump(ptr, len);
-            ret = write((int)conn, ptr, len);
+            ret = write((int)(intptr_t)conn, ptr, len);
             if (ret < 0)
                 perror("send");
         } while ((ret < 0) && (errno == EAGAIN || errno == EINTR));
@@ -445,7 +445,7 @@ static OSStatus SocketRead(SSLConnectionRef conn, void *data, size_t *length)
     do {
         ssize_t ret;
         do {
-            ret = read((int)conn, ptr, len);
+            ret = read((int)(intptr_t)conn, ptr, len);
             if (ret < 0)
                 perror("send");
         } while ((ret < 0) && (errno == EAGAIN || errno == EINTR));
