+++ Security/dtlsEcho/dtlsEchoServer.c
@@ -182,7 +182,7 @@ OSStatus SocketRead(
                     void 				*data,
                     size_t 				*dataLength)
 {
-    int fd = (int)connection;
+    int fd = (int)(intptr_t)connection;
     ssize_t len;
     uint8_t *d=readBuffer;
 
@@ -230,7 +230,7 @@ OSStatus SocketWrite(
                      const void         *data,
                      size_t 			*dataLength)	/* IN/OUT */
 {
-    int fd = (int)connection;
+    int fd = (int)(intptr_t)connection;
     ssize_t len;
     OSStatus err = errSecSuccess;
     const uint8_t *d=data;
